combined types files
