+++ client/src/react-app-env.d.ts
@@ -1,3 +1,17 @@
+declare global {
+  interface Window {
+    SpeechRecognition: SpeechRecognition;
+    webkitSpeechRecognition: SpeechRecognition;
+  }
+}
+
+declare module "cloudinary-react" {
+  import { Component } from "react";
+
+  export class CloudinaryContext extends Component<any, any> {}
+  export class Video extends Component<any, any> {}
+}
+
 /// <reference types="react-scripts" />
 
 //general types
