--- client/src/react-cloudinary.d.ts
@@ -1,6 +1,0 @@
-declare module "cloudinary-react" {
-  import { Component } from "react";
-
-  export class CloudinaryContext extends Component<any, any> {}
-  export class Video extends Component<any, any> {}
-}
