--- client/src/speech-recognition.d.ts
@@ -1,8 +1,0 @@
-declare global {
-  interface Window {
-    SpeechRecognition: SpeechRecognition;
-    webkitSpeechRecognition: SpeechRecognition;
-  }
-}
-
-export {};
